Captive Portal settings: Re-add 'GrapheneOS'
When updating from my LineageOS 19.1 build, the settings app has crashed,
if the previous default "GrapheneOS" was active - the array entry was
commented out in the originating DOS patch, but the controller code has
recognized the setting and assigned array index 11 to it, which has lead
to an ArrayIndexOutOfBoundsException.

Further, entirely remove XIAOMI from the code.

Change-Id: Ie3d2c2924412ef225cb3481cfee40f56bff70976